requirements.txt
annoy==1.17.3
codetiming==1.4.0
matplotlib==3.10.3
networkx==3.4.2
noise==1.2.2
numba==0.61.2
numpy
opencv_python==4.11.0.86
Pillow==11.3.0
plotly==6.1.2
scikit_learn==1.7.1
scipy==1.16.0
